Removing Docs from mainline
This branch removes the docs completely from the mainline branch. The
rationale behind this is we are leveraging GitHub pages for the
documentation of this project. Instead of having documentation in two
places and a convoluted documentation system, we are opting to house the
docs in an orphan branch -per the recommended procedure on GitHub- and
all documentation efforts need to be edited then PR approved to be
merged into the `gh-pages` branch. Indeed this will require a little
different workflow but in the end, I think this will greatly simplify
things and it should also prevent wasted effort. An added bonus is it
keeps the mainline a little cleaner which is kinda nice. Questions and
comments welcome.