Update package.json
Package.json is missing a main entry so I can't use it in a project. Also, if you list Vue as a peerDependency rather than a dependency it won't install two different versions of Vue in my project. But I added as a devDependency so it installs for you for development